Address comments

+++ .kokoro/tests/run_single_test.sh
@@ -43,10 +43,14 @@ EXIT=$?
 
 # Inject region tag data into the test log
 XUNIT_PATH="$PWD/sponge_log.xml"
-XUNIT_TMP_PATH="$PWD/drift_tmp.xml"
+XUNIT_TMP_PATH="tmp/drift_tmp.xml"
 
-cp $XUNIT_PATH $XUNIT_TMP_PATH
-cat "$XUNIT_TMP_PATH" | python3.8 "$PARSER_PATH" inject-snippet-mapping "$PWD" > "$XUNIT_PATH"
+if [ -f "$XUNIT_PATH" ]; then
+    cp $XUNIT_PATH $XUNIT_TMP_PATH
+    cat "$XUNIT_TMP_PATH" | python3.8 "$PARSER_PATH" inject-snippet-mapping "$PWD" > "$XUNIT_PATH"
+else
+    echo "No XUnit output file found!"
+fi
 
 # If REPORT_TO_BUILD_COP_BOT is set to "true", send the test log
 # to the Build Cop Bot.
